+++ sample-notificator-app/src/main/groovy/life/qbic/samplenotificator/DependencyManager.groovy
@@ -1,0 +1,9 @@
+package life.qbic.samplenotificator
+/**
+ * <h1>Sets up the use cases</h1>
+ *
+ * @since 1.0.0
+ *
+*/
+class DependencyManager {
+}
