+++ sample-notificator-domain/src/main/groovy/life/qbic/business/notification/create/CreateNotificationInput.groovy
@@ -1,11 +1,11 @@
-package life.qbic.business.notification.send
+package life.qbic.business.notification.create
 /**
- * <h1>Interface to access and trigger the {@link SendNotification} use case</h1>
+ * <h1>Interface to access and trigger the {@link CreateNotification} use case</h1>
  *
  * @since 1.0.0
  *
 */
-interface SendNotificationInput {
+interface CreateNotificationInput {
 
     /**
      * Trigger sending notifications to a list of subscribers about changes in their subscriptions
