+++ sample-notificator-domain/src/main/groovy/life/qbic/business/notification/create/CreateNotificationOutput.groovy
@@ -1,0 +1,13 @@
+package life.qbic.business.notification.create
+
+/**
+ * <h1><short description></h1>
+ *
+ * <p><detailed description></p>
+ *
+ * @since 1.0.0
+ *
+*/
+interface CreateNotificationOutput {
+
+}
