+++ sample-notificator-domain/src/main/groovy/life/qbic/business/notification/send/SendNotificationOutput.groovy
@@ -7,7 +7,6 @@ package life.qbic.business.notification.send
  * @since <versiontag>
  *
 */
-class SendNotification {
-
-    //todo add logic to send the created notifcations via email
+interface SendNotificationOutput {
+    //todo add methods
 }
